+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Add `Class` enum
 - Achievement : Add `set_unlock_skin` (TODO)
 - Sprite : Add `.width`, `.height`, and `.subimages` getters to wrapper
+- Callback : Add `add_SO`
 
 ### v0.1.43
 - Survivor : Set `pHmax_base` to `2.8` on init